Docker changes

+++ db/setup.py
@@ -1,12 +1,14 @@
 #Once the database is setup, this will create all the tables defined in models
 import os
-from sqlalchemy import create_engine
-from models import *
+from sqlalchemy import create_engine, inspect
+from .models import *
 
 def build():
     engine = create_engine(os.getenv("DATABASE_URL"))
+    inspector = inspect(engine)
 
-    Base.metadata.drop_all(engine)
-    Base.metadata.create_all(engine)
+    if 'drone_info' not in inspector.get_table_names():
+        Base.metadata.drop_all(engine)
+        Base.metadata.create_all(engine)
 
 build()
